created hook to get swap to amount
